Refactor hunt page and improve alert visibility
Refactored `DashboardHuntPage` to dynamically fetch hunt data,
conditionally render content based on hunt status, and add
polling for updates. Introduced `DashboardActiveHunt`
component to encapsulate active hunt UI, reintroducing
player rankings and item carousel. Enhanced `Alert.jsx`
styling with `zIndex: 1000` for better visibility.
